Move auth constants to keep relevant code colocated
While still supporting Cypress tests
