+++ apps/central-scan/integration-testing/cypress/e2e/configuration.cy.ts
@@ -2,15 +2,14 @@ import { Buffer } from 'buffer';
 import { methodUrl } from '@votingworks/grout';
 import { sha256 } from 'js-sha256';
 
-// Importing all of @votingworks/auth causes Cypress tests to fail since Cypress doesn't seem to
-// interact well with PCSC Lite card reader code
-// eslint-disable-next-line vx/no-import-workspace-subfolders
-import { DEV_JURISDICTION } from '@votingworks/auth/src/constants';
+// Importing all of @votingworks/auth causes Cypress tests to fail since @votingworks/auth contains
+// code that isn't browser-safe
 // eslint-disable-next-line vx/no-import-workspace-subfolders
 import {
+  DEV_JURISDICTION,
   mockCard,
   MockFileContents,
-} from '@votingworks/auth/src/mock_file_card';
+} from '@votingworks/auth/src/cypress';
 
 const PIN = '000000';
