+++ apps/mark/integration-testing/cypress/support/e2e.js
@@ -19,12 +19,13 @@ import './commands';
 import { electionSampleDefinition } from '@votingworks/fixtures';
 import { methodUrl } from '@votingworks/grout';
 
-// Importing all of @votingworks/auth causes Cypress tests to fail since Cypress doesn't seem to
-// interact well with PCSC Lite card reader code
+// Importing all of @votingworks/auth causes Cypress tests to fail since @votingworks/auth contains
+// code that isn't browser-safe
 // eslint-disable-next-line vx/no-import-workspace-subfolders
-import { DEV_JURISDICTION } from '@votingworks/auth/src/constants';
-// eslint-disable-next-line vx/no-import-workspace-subfolders
-import { mockCard } from '@votingworks/auth/src/mock_file_card';
+import {
+  DEV_JURISDICTION,
+  mockCard,
+} from '@votingworks/auth/src/cypress';
 
 const { electionData, electionHash } = electionSampleDefinition;
 const PIN = '000000';
