+++ libs/auth/scripts/generate_dev_keys_and_certs.ts
@@ -6,11 +6,15 @@ import { electionFamousNames2021Fixtures } from '@votingworks/fixtures';
 
 import { CardDetails } from '../src/card';
 import {
+  CardType,
+  CERT_EXPIRY_IN_DAYS,
   constructCardCertSubject,
   constructCardCertSubjectWithoutJurisdictionAndCardType,
   constructMachineCertSubject,
+  STANDARD_CERT_FIELDS,
 } from '../src/certs';
 import { DEV_PRIVATE_KEY_PASSWORD } from '../src/java_card_config';
+import { DEV_JURISDICTION } from '../src/jurisdictions';
 import {
   certPemToDer,
   createCert,
@@ -19,12 +23,6 @@ import {
   publicKeyPemToDer,
 } from '../src/openssl';
 import { runCommand } from './utils';
-import {
-  CERT_EXPIRY_IN_DAYS,
-  DEV_JURISDICTION,
-  STANDARD_CERT_FIELDS,
-} from '../src/constants';
-import { CardType } from '../src/types';
 
 async function generateDevPrivateKey(): Promise<Buffer> {
   const privateKey = await openssl([
