+++ libs/auth/src/index.ts
@@ -1,10 +1,10 @@
-export { DEV_JURISDICTION } from './constants';
 export * from './dipped_smart_card_auth_api';
 export * from './dipped_smart_card_auth';
 export * from './inserted_smart_card_auth_api';
 export * from './inserted_smart_card_auth';
 export * from './java_card_config';
 export * from './java_card';
+export * from './jurisdictions';
 export * from './memory_card';
 export * from './mock_file_card';
 export * from './test_utils';
