+++ libs/auth/src/jurisdictions.ts
@@ -1,0 +1,4 @@
+/**
+ * The jurisdiction in all dev certs
+ */
+export const DEV_JURISDICTION = 'st.dev-jurisdiction';
